lib/list.c: Use size_t for array indices and sizes
Reported-by: Tobias Stoeckmann <tobias@stoeckmann.org>
Suggested-by: Tobias Stoeckmann <tobias@stoeckmann.org>

+++ lib/list.c
@@ -8,9 +8,7 @@
 
 #include "config.h"
 
-#ident "$Id$"
-
-#include <assert.h>
+#include <stddef.h>
 
 #include "alloc/malloc.h"
 #include "prototypes.h"
@@ -19,6 +17,8 @@
 #include "string/strcmp/streq.h"
 #include "string/strdup/strdup.h"
 #include "string/strtok/strsep2ls.h"
+
+#include <assert.h>
 
 
 /*
@@ -31,8 +31,8 @@
 /*@only@*/char **
 add_list(/*@returned@*/ /*@only@*/char **list, const char *member)
 {
-	int i;
-	char **tmp;
+	size_t  i;
+	char    **tmp;
 
 	assert (NULL != member);
 	assert (NULL != list);
@@ -82,8 +82,8 @@ add_list(/*@returned@*/ /*@only@*/char **list, const char *member)
 /*@only@*/char **
 del_list(/*@returned@*/ /*@only@*/char **list, const char *member)
 {
-	int i, j;
-	char **tmp;
+	size_t  i, j;
+	char    **tmp;
 
 	assert (NULL != member);
 	assert (NULL != list);
@@ -137,8 +137,8 @@ del_list(/*@returned@*/ /*@only@*/char **list, const char *member)
 /*@only@*/char **
 dup_list(char *const *list)
 {
-	int i;
-	char **tmp;
+	size_t  i;
+	char    **tmp;
 
 	assert (NULL != list);
